modularized panelApp by separating components and created types file
