+++ src/panel/components/navBar.tsx
@@ -3,9 +3,7 @@ import AppBar from '@mui/material/AppBar';
 import Box from '@mui/material/Box';
 import Toolbar from '@mui/material/Toolbar';
 import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
 import IconButton from '@mui/material/IconButton';
-import MenuIcon from '@mui/icons-material/Menu';
 import FlashlightOnIcon from '@mui/icons-material/FlashlightOn';
 import Skeleton from '@mui/material/Skeleton';
